move components to correct folder
