+++ mdx-components.tsx
@@ -1,6 +1,6 @@
 import { type MDXComponents as MDXComponentsType } from 'mdx/types'
 
-import { MDXComponents } from '@/components/MDXComponents'
+import { MDXComponents } from '@/components/ui/MDXComponents'
 
 export function useMDXComponents(components: MDXComponentsType) {
   return {
