+++ src/app/[locale]/blog/[id]/ViewBlog.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { Border } from '@/components/Border'
-import { FadeIn } from '@/components/FadeIn'
-import { MessageMarkdown } from '@/components/message-markdown'
+import { FadeIn } from '@/components/ui/FadeIn'
+import { MessageMarkdown } from '@/components/ui/message-markdown'
 import { PageLinks } from '@/components/PageLinks'
 import { fetchAxiosAPI } from '@/request/request'
 import { useEffect, useState } from 'react'
