+++ src/app/[locale]/contact/page.tsx
@@ -3,10 +3,10 @@ import Link from 'next/link'
 import { useId } from 'react'
 
 import { Border } from '@/components/Border'
-import { Button } from '@/components/Button'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
-import { Offices } from '@/components/Offices'
+import { Button } from '@/components/elements/Button'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
+import { Offices } from '@/components/sections/Offices'
 import { PageIntro } from '@/components/sections/PageIntro'
 import { SocialMedia } from '@/components/SocialMedia'
 import { fetchAxiosAPI } from '@/request/request'
