+++ src/app/[locale]/projects/ViewProjects.tsx
@@ -1,6 +1,6 @@
-import Clients from '@/components/Clients'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
+import Clients from '@/components/sections/Clients'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
 import ProjectCard from '@/components/ProjectCard'
 import { ContactSection } from '@/components/sections/ContactSection'
 import { PageIntro } from '@/components/sections/PageIntro'
