+++ src/app/[locale]/projects/[id]/ViewProject.tsx
@@ -1,9 +1,9 @@
-import BasicMarkdown from '@/components/BasicMarkdown'
+import BasicMarkdown from '@/components/ui/BasicMarkdown'
 import { Testimonial } from '@/components/Testimonial'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
-import { GrayscaleTransitionImage } from '@/components/GrayscaleTransitionImage'
-import { MessageMarkdown } from '@/components/message-markdown'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
+import { GrayscaleTransitionImage } from '@/components/ui/GrayscaleTransitionImage'
+import { MessageMarkdown } from '@/components/ui/message-markdown'
 import { PageIntro } from '@/components/sections/PageIntro'
 import { fetchAxiosAPI } from '@/request/request'
 import { RestQueryParams } from '@/types/global'
