+++ src/app/[locale]/projects/oldExample.tsx
@@ -4,10 +4,10 @@ import Link from 'next/link'
 
 import { Blockquote } from '@/components/Blockquote'
 import { Border } from '@/components/Border'
-import { Button } from '@/components/Button'
+import { Button } from '@/components/elements/Button'
 import { ContactSection } from '@/components/sections/ContactSection'
-import { Container } from '@/components/Container'
-import { FadeIn, FadeInStagger } from '@/components/FadeIn'
+import { Container } from '@/components/ui/Container'
+import { FadeIn, FadeInStagger } from '@/components/ui/FadeIn'
 import { PageIntro } from '@/components/sections/PageIntro'
 import { Testimonial } from '@/components/Testimonial'
 import logoBrightPath from '@/images/clients/bright-path/logo-dark.svg'
