+++ src/app/[locale]/projects/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
-import Clients from '@/components/Clients'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
+import Clients from '@/components/sections/Clients'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
 import ProjectCard from '@/components/ProjectCard'
 import { ContactSection } from '@/components/sections/ContactSection'
 import { PageIntro } from '@/components/sections/PageIntro'
