+++ src/app/[locale]/services/[id]/ViewService.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { Border } from '@/components/Border'
-import { FadeIn } from '@/components/FadeIn'
+import { FadeIn } from '@/components/ui/FadeIn'
 import { fetchAxiosAPI } from '@/request/request'
 import { BaseData } from '@/types/global'
 import ReactMarkdown from 'react-markdown'
