+++ src/app/[locale]/services/ai/page.tsx
@@ -1,20 +1,18 @@
 import { type Metadata } from 'next'
 
-import { Blockquote } from '@/components/Blockquote'
 import { ContactSection } from '@/components/sections/ContactSection'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
-import { GridList, GridListItem } from '@/components/GridList'
-import { GridPattern } from '@/components/GridPattern'
-import { List, ListItem } from '@/components/List'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
+import { GridList, GridListItem } from '@/components/ui/GridList'
+import { GridPattern } from '@/components/ui/GridPattern'
+import { List, ListItem } from '@/components/ui/List'
 import { PageIntro } from '@/components/sections/PageIntro'
-import { SectionIntro } from '@/components/SectionIntro'
+import { SectionIntro } from '@/components/sections/SectionIntro'
 import { StylizedImage } from '@/components/StylizedImage'
 import { TagList, TagListItem } from '@/components/TagList'
 import imageLaptop from '@/images/laptop.jpg'
 import imageMeeting from '@/images/meeting.jpg'
 import imageWhiteboard from '@/images/whiteboard.jpg'
-import Features from '@/components/Features'
 import FeatureDetail from '@/components/FeatureDetail'
 
 function Section({
