+++ src/app/[locale]/work/wrapper.tsx
@@ -1,8 +1,8 @@
 import { ContactSection } from '@/components/sections/ContactSection'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
-import { GrayscaleTransitionImage } from '@/components/GrayscaleTransitionImage'
-import { MDXComponents } from '@/components/MDXComponents'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
+import { GrayscaleTransitionImage } from '@/components/ui/GrayscaleTransitionImage'
+import { MDXComponents } from '@/components/ui/MDXComponents'
 import { PageIntro } from '@/components/sections/PageIntro'
 import { PageLinks } from '@/components/PageLinks'
 import { type CaseStudy, type MDXEntry, loadCaseStudies } from '@/lib/mdx'
