+++ src/components/elements/FeatureDetail.tsx
