+++ src/components/elements/PostCard.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
 import { Button } from './Button'
-import { Border } from './Border'
+import { Border } from '../ui/Border'
 import { formatDate } from '@/lib/formatDate'
 import ReactMarkdown from 'react-markdown'
 import { PageIntro } from '@/types/global'
