+++ src/components/elements/WorkCard.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { FadeIn } from '@/components/FadeIn'
+import { FadeIn } from '@/components/ui/FadeIn'
 import Image from 'next/image'
 import Link from 'next/link'
 import { Blockquote } from '@/components/Blockquote'
 import { Border } from '@/components/Border'
 
-import { Button } from '@/components/Button'
+import { Button } from '@/components/elements/Button'
 import { PageIntro } from '@/types/global'
 
 interface WorkProps {
