+++ src/components/images/StylizedImage.tsx
