+++ src/components/sections/Posts.tsx
@@ -2,11 +2,11 @@ import { type Metadata } from 'next'
 import Image from 'next/image'
 import Link from 'next/link'
 
-import { Border } from '@/components/Border'
-import { Button } from '@/components/Button'
+import { Border } from '@/components/ui/Border'
+import { Button } from '@/components/elements/Button'
 import { ContactSection } from '@/components/sections/ContactSection'
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
 import { PageIntro } from '@/components/sections/PageIntro'
 import { formatDate } from '@/lib/formatDate'
 import { loadArticles } from '@/lib/mdx'
