+++ src/components/sections/ProjectsSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { SectionIntro } from '../SectionIntro'
-import { Container } from '../Container'
-import { FadeIn, FadeInStagger } from '../FadeIn'
+import { SectionIntro } from './SectionIntro'
+import { Container } from '../ui/Container'
+import { FadeIn, FadeInStagger } from '../ui/FadeIn'
 import Link from 'next/link'
 import ReactMarkdown from 'react-markdown'
 import { Project } from '@/types/project'
