+++ src/components/sections/SectionIntro.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx'
 
-import { Container } from '@/components/Container'
-import { FadeIn } from '@/components/FadeIn'
+import { Container } from '@/components/ui/Container'
+import { FadeIn } from '@/components/ui/FadeIn'
 
 export function SectionIntro({
   title,
