+++ src/components/ui/GrayscaleTransitionImage.tsx
